+++ src/DataTypes/Type/NumericType.php
@@ -29,6 +29,12 @@ class NumericType extends AbstractType
      */
     public function fromSchema(mixed $value, ?ExpectedType $expected = null): mixed
     {
+        if (null === $value) {
+            $this->assertNullable($expected);
+
+            return null;
+        }
+
         $this->assertScalar($value);
 
         if (null !== $expected) {
@@ -51,6 +57,10 @@ class NumericType extends AbstractType
      */
     public function toSchema(mixed $value, ?ExpectedType $expected = null): mixed
     {
+        if (null === $value) {
+            return null;
+        }
+
         $this->assertScalar($value);
 
         settype($value, $this->numericType);